requirements.txt
awscli==1.14.36
boto3==1.5.26
botocore==1.8.40
certifi==2017.11.5
chardet==3.0.4
click==6.7
colorama==0.3.7
docutils==0.14
enum-compat==0.0.2
enum34==1.1.6
eventlet==0.21.0
Flask==0.12.2
Flask-Login==0.4.0
Flask-PyMongo==0.5.1
Flask-SocketIO==2.9.2
future==0.16.0
greenlet==0.4.12
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
numpy==1.13.3
oauthlib==2.0.6
pandas==0.21.0
pyasn1==0.4.2
pyhottop
pymongo==3.5.1
pyserial==3.4
python-dateutil==2.6.1
python-engineio==2.0.1
python-socketio==1.8.3
python-twitter==3.3
pytz==2017.3
PyYAML==3.12
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
s3transfer==0.1.12
scipy==1.0.0
six==1.11.0
urllib3==1.22
Werkzeug==0.12.2
WTForms==2.1
